pyproject.toml
[project]
name = "apiris"
version = "1.0.0"
description = "Apiris - Deterministic AI Reliability Intelligence SDK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pyyaml==6.0.2",
  "requests==2.32.3",
  "fastapi==0.115.0",
  "uvicorn==0.30.1",
  "rich==13.7.0",
  "typer>=0.15.1",
]

[project.scripts]
apiris = "apiris.cli:main"

[project.optional-dependencies]
test = [
  "pytest==8.3.4",
  "responses==0.25.3",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["apiris*"]
